posts_service/requirements.txt
grpcio-tools==1.71.0
grpcio==1.71.0
googleapis-common-protos==1.69.2
kafka-python==2.0.2
